Split gatsby action
